+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Entities spawned during component deserialization now also get the `Remote` marker.
+
 ## [0.39.1] - 2026-03-09
 
 ### Fixed