+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Replicated` is no longer automatically inserted on clients, only `Remote`. `scene::replicate_into` will serialize all entities that have either `Remote` or `Replicated`.
 
+## [0.39.2] - 2026-04-01
+
 ### Fixed
 
 - Entities spawned during component deserialization now also get the `Remote` marker.
@@ -1041,7 +1043,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release after separation from [Project Harmonia](https://github.com/simgine/project_harmonia).
 
-[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.39.1...HEAD
+[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.39.2...HEAD
+[0.39.2]: https://github.com/simgine/bevy_replicon/compare/v0.39.1...v0.39.2
 [0.39.1]: https://github.com/simgine/bevy_replicon/compare/v0.39.0...v0.39.1
 [0.39.0]: https://github.com/simgine/bevy_replicon/compare/v0.38.2...v0.39.0
 [0.38.2]: https://github.com/simgine/bevy_replicon/compare/v0.38.1...v0.38.2